--- tests/e2e/https_proxy/config/host.yaml
@@ -1,14 +1,0 @@
-# yaml-language-server: $schema=https://json.schemastore.org/traefik-v3.json
-http:
-  routers:
-    default-router:
-      rule: 'Host(`{{ env "HOST" }}`)'
-      priority: 10000
-      service: app
-      entrypoints:
-        - websecure
-  services:
-    app:
-      loadBalancer:
-        servers:
-          - url: http://{{ env "HOST" }}:{{ env "PORT" }}
